Remove About and Design Cleanup
- Updated the designs and removed about screen altogether.
